Update serial_pm.py
In Home Assistant Particulate Matter result was inconsistent in time. It was because serial device is open once, so data from sensor was collected by system in device buffer and slowly read from it. So results was "extended" in time. So added here is cleaning input buffer before next read of PM value. Additionally during research was added check sum control for PLANTOWER1 config.