Add checksum verification.
A bad checksum should raise an exception but we stay consistent with the rest of the code and return None if the checksum fails.

Note: My just-plugged in MH-Z19B sensor will return, in order:
- A bad checksum
- An incorrect PPM value of 1215, twice
- 410 PPM for approximately one more minute until the device has warmed up.

As this code doesn't know its history it's up to the calling function to keep track and potentially blacklist these values.